+++ subsys/usb/device/class/msc.c
@@ -39,12 +39,71 @@
 #include <zephyr/sys/byteorder.h>
 #include <zephyr/sys/__assert.h>
 #include <zephyr/storage/disk_access.h>
-#include <zephyr/usb/class/usb_msc.h>
 #include <zephyr/usb/usb_device.h>
 #include <usb_descriptor.h>
 
 #include <zephyr/logging/log.h>
 LOG_MODULE_REGISTER(usb_msc, CONFIG_USB_MASS_STORAGE_LOG_LEVEL);
+
+/* MSC Subclass and Protocol Codes */
+#define SCSI_TRANSPARENT_SUBCLASS	0x06
+#define BULK_ONLY_TRANSPORT_PROTOCOL	0x50
+
+/* MSC Request Codes for Bulk-Only Transport */
+#define MSC_REQUEST_GET_MAX_LUN		0xFE
+#define MSC_REQUEST_RESET		0xFF
+
+/* MSC Command Block Wrapper (CBW) Signature */
+#define CBW_Signature			0x43425355
+
+/* MSC Command Block Wrapper Flags */
+#define CBW_DIRECTION_DATA_IN		0x80
+
+/* MSC Bulk-Only Command Block Wrapper (CBW) */
+struct CBW {
+	uint32_t Signature;
+	uint32_t Tag;
+	uint32_t DataLength;
+	uint8_t  Flags;
+	uint8_t  LUN;
+	uint8_t  CBLength;
+	uint8_t  CB[16];
+} __packed;
+
+/* MSC Command Status Wrapper (CBW) Signature */
+#define CSW_Signature			0x53425355
+
+/* MSC Command Block Status Values */
+#define CSW_STATUS_CMD_PASSED		0x00
+#define CSW_STATUS_CMD_FAILED		0x01
+#define CSW_STATUS_PHASE_ERROR		0x02
+
+/* MSC Bulk-Only Command Status Wrapper (CSW) */
+struct CSW {
+	uint32_t Signature;
+	uint32_t Tag;
+	uint32_t DataResidue;
+	uint8_t  Status;
+} __packed;
+
+/* SCSI transparent command set used by MSC */
+#define TEST_UNIT_READY			0x00
+#define REQUEST_SENSE			0x03
+#define FORMAT_UNIT			0x04
+#define INQUIRY				0x12
+#define MODE_SELECT6			0x15
+#define MODE_SENSE6			0x1A
+#define START_STOP_UNIT			0x1B
+#define MEDIA_REMOVAL			0x1E
+#define READ_FORMAT_CAPACITIES		0x23
+#define READ_CAPACITY			0x25
+#define READ10				0x28
+#define WRITE10				0x2A
+#define VERIFY10			0x2F
+#define READ12				0xA8
+#define WRITE12				0xAA
+#define MODE_SELECT10			0x55
+#define MODE_SENSE10			0x5A
 
 /* max USB packet size */
 #define MAX_PACKET	CONFIG_MASS_STORAGE_BULK_EP_MPS
